FSD architecture is applied
