+++ package.json
@@ -26,11 +26,11 @@
     "@heroui/use-theme": "2.1.8",
     "@react-aria/visually-hidden": "3.8.22",
     "@react-types/shared": "3.29.0",
+    "axios": "^1.9.0",
     "clsx": "2.1.1",
     "framer-motion": "11.15.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-hook-form": "^7.56.3",
     "react-router-dom": "6.23.0",
     "tailwind-variants": "0.3.0",
     "tailwindcss": "3.4.16"
