+++ src/app/App.tsx
@@ -1,5 +1,5 @@
+import MainPage from '@/pages/MainPage/MainPage';
 import { FC } from 'react';
-import MainPage from './MainPage/MainPage';
 
 const App: FC = () => <MainPage />;
