+++ src/main.tsx
@@ -1,16 +1,13 @@
-import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import './styles/globals.css';
+import './app/styles/globals.css';
 import { HeroUIProvider } from '@heroui/system';
-import App from './App';
+import App from './app/App';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <HeroUIProvider>
-        <App />
-      </HeroUIProvider>
-    </BrowserRouter>
-  </React.StrictMode>,
+  <BrowserRouter>
+    <HeroUIProvider>
+      <App />
+    </HeroUIProvider>
+  </BrowserRouter>,
 );
