+++ src/shared/ui/hero/primitives.ts
