+++ tsconfig.json
@@ -24,6 +24,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src", "public/assets"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
